responsive navbar added

+++ src/pages/Card/DonateCard.jsx
@@ -1,52 +1,57 @@
 import PropTypes from 'prop-types';
+import swal from 'sweetalert';
 
 const DonateCard = ({ cardOne }) => {
 
-    const { id, picture, title, category, price, text_button_bg_color, card_bg_color } = cardOne || {}
+    const { id, picture, title, price, description, benefits, text_button_bg_color } = cardOne || {}
 
     const addedDonationArray = [];
     console.log("image show: ", id);
     const handleAddDonation = () => {
-        // console.log(cardOne);
+
         const donatedItems = JSON.parse(localStorage.getItem("donation"));
+
+
         if (!donatedItems) {
             addedDonationArray.push(cardOne);
             localStorage.setItem("donation", JSON.stringify(addedDonationArray));
+            swal("Good job!", "Donation added successfully!", "success");
         }
         else {
-            addedDonationArray.push(...donatedItems, cardOne);
-            localStorage.setItem("donation", JSON.stringify(addedDonationArray));
+            const isExists = donatedItems.find(cardOne => cardOne.id == id)
+            if (!isExists) {
+                addedDonationArray.push(...donatedItems, cardOne);
+                localStorage.setItem("donation", JSON.stringify(addedDonationArray));
+                swal("Good job!", "Donation added successfully!", "success");
+            }
+            else {
+                swal("Alert!", "Already added", "warning");
+            }
         }
     }
 
     return (
-        <div className="flex justify-center items-center h[80vh]">
-            <div className="flex w-full max-w-[44rem] flex-row bg-white shadow-md rounded-xl" style={{ color: text_button_bg_color, backgroundColor: card_bg_color }}>
-                <div className=" m-0 w-2/5 overflow-hidden rounded-xl">
-                    <img className="h-full w-full object-cover" src={picture} alt="image" />
-                </div>
-
-                <div className="p-6">
-                    <h6 className="mb-4 w-[7rem] rounded-md block text-base font-semibold uppercase text-center"
-                        style={{ color: text_button_bg_color, backgroundColor: card_bg_color }}>
-                        {category}
-                    </h6>
-                    <h4 className="mb-2 text-2xl font-semibold text-black">
-                        {title}
-                    </h4>
-                    <p className="text-xl font-semibold">${price}</p>
-                    <a className="inline-block" href="#">
-                        <button
-                            onClick={handleAddDonation}
-                            className="flex items-center gap-2 rounded-lg mt-5 py-2 px-6 text-center align-middle text-sm font-bold uppercase transition-all disabled:pointer-events-none disabled:opacity-50 disabled:shadow-none text-white"
-                            type="button" style={{ backgroundColor: text_button_bg_color }}
-                        >
-                            View Details
-                        </button>
-                    </a>
-                </div>
+        <div>
+            <div>
+                <img className='w-full h-[70vh] object-cover z-10' src={picture} alt="" />
             </div>
-        </div>
+            <div className='relative'>
+                <div className='absolute z-20 w-full mt-[-100px] h-[100px] opacity-50 bg-black'></div>
+                <button style={{ backgroundColor: text_button_bg_color }} onClick={handleAddDonation} className='absolute z-30 ml-10 mt-[-70px] bg-red-500 text-white px-6 py-2 rounded-md font-semibold'> <span>Donate ${price}</span> </button>
+            </div>
+            <h3 className='text-3xl font-bold mt-5'>{title}</h3>
+            <p className='text-md mt-2 mb-5'>{description} </p>
+
+            <div>
+                <h3 className='text-xl font-bold'>Here are some of the benefits of a good education:</h3>
+                <ul className='list-disc list-inside ml-4'>
+                    {
+                        benefits?.map(benefit => <li key={benefit.id}>{benefit}</li>)
+                    }
+                </ul>
+            </div>
+
+        </div >
 
     );
 };
